src/eCommerce_application/requirements.txt
bandit==1.8.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
exceptiongroup==1.2.2
Flask==3.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pbr==6.1.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.4
shellingham==1.5.4
stevedore==5.4.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
